+++ src/main/java/kitchenpos/table/bo/TableBo.java
@@ -1,9 +1,9 @@
-package kitchenpos.bo;
+package kitchenpos.table.bo;
 
-import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.model.OrderStatus;
-import kitchenpos.model.OrderTable;
+import kitchenpos.order.dao.OrderDao;
+import kitchenpos.order.dao.OrderTableDao;
+import kitchenpos.order.model.OrderStatus;
+import kitchenpos.table.model.OrderTable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
