+++ src/test/java/kitchenpos/table/TableGroupBoTest.java
@@ -1,9 +1,12 @@
-package kitchenpos.bo;
+package kitchenpos.table;
 
-import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
-import kitchenpos.model.TableGroup;
+import kitchenpos.order.InMemoryOrderDao;
+import kitchenpos.order.InMemoryOrderTableDao;
+import kitchenpos.order.dao.OrderDao;
+import kitchenpos.order.dao.OrderTableDao;
+import kitchenpos.table.dao.TableGroupDao;
+import kitchenpos.table.model.TableGroup;
+import kitchenpos.table.bo.TableGroupBo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
