+++ src/test/java/kitchenpos/table/controller/TableGroupRestControllerTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.controller;
+package kitchenpos.table.controller;
 
-import kitchenpos.bo.TableGroupBo;
-import kitchenpos.model.TableGroup;
+import kitchenpos.table.bo.TableGroupBo;
+import kitchenpos.table.model.TableGroup;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
